refact: Reorganize code
